hardened_agent/requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# Utilities
pydantic>=2.0.0
